Rename component to SearchFor, add route and setState in component
