+++ src/components/Navbar/NavBar.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import './Navbar.css';
-import SearchMovie from '../searchMovie/SearchMovie';
+import SearchMovie from '../SearchFor/SearchFor';
 
 const Navbar = () => {
 	// TODO: Set navbar to fixed postion on scroll,
